% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penn115_fname.R
\name{penn115_fname}
\alias{penn115_fname}
\alias{penn115_image_fname}
\alias{penn115_brain_fname}
\alias{penn115_brain_mask_fname}
\alias{penn115_brain_mask_probability_fname}
\alias{penn115_registration_mask_fname}
\alias{penn115_nifti}
\alias{penn115_image}
\alias{penn115_brain}
\alias{penn115_brain_mask}
\alias{penn115_brain_mask_probability}
\alias{penn115_registration_mask}
\title{Filename of penn115 image}
\usage{
penn115_fname(type = c("brain", "image", "brain_mask",
  "brain_mask_probability", "registration_mask"))

penn115_image_fname()

penn115_brain_fname()

penn115_brain_mask_fname()

penn115_brain_mask_probability_fname()

penn115_registration_mask_fname()

penn115_nifti(...)

penn115_image()

penn115_brain()

penn115_brain_mask()

penn115_brain_mask_probability()

penn115_registration_mask()
}
\arguments{
\item{type}{type of image to return}

\item{...}{arguments to pass to \code{\link{penn115_fname}}}
}
\value{
A character vector
}
\description{
Filename of penn115 image
}
\examples{
penn115_fname(type = "image")
penn115_fname(type = "brain")
penn115_fname(type = "brain_mask")
penn115_fname(type = "brain_mask_probability")
penn115_fname(type = "registration_mask")

img = penn115_image()

}
