% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penn60_fname.R
\name{penn60_fname}
\alias{penn60_fname}
\alias{penn60_image_fname}
\alias{penn60_brain_fname}
\alias{penn60_brain_mask_fname}
\alias{penn60_brain_mask_probability_fname}
\alias{penn60_registration_mask_fname}
\alias{penn60_nifti}
\alias{penn60_image}
\alias{penn60_brain}
\alias{penn60_brain_mask}
\alias{penn60_brain_mask_probability}
\alias{penn60_registration_mask}
\title{Filename of penn60 image}
\usage{
penn60_fname(type = c("brain", "image", "brain_mask",
  "brain_mask_probability", "registration_mask"))

penn60_image_fname()

penn60_brain_fname()

penn60_brain_mask_fname()

penn60_brain_mask_probability_fname()

penn60_registration_mask_fname()

penn60_nifti(...)

penn60_image()

penn60_brain()

penn60_brain_mask()

penn60_brain_mask_probability()

penn60_registration_mask()
}
\arguments{
\item{type}{type of image to return}

\item{...}{arguments to pass to \code{\link{penn60_fname}}}
}
\value{
A character vector
}
\description{
Filename of penn60 image
}
\examples{
penn60_fname(type = "image")
penn60_fname(type = "brain")
penn60_fname(type = "brain_mask")
penn60_fname(type = "brain_mask_probability")
penn60_fname(type = "registration_mask")

img = penn60_image()

}
